- Add logger to engine - Add build config to project and engine - Create basics of draw-on-touch system

+++ app/src/main/java/ru/acediat/simplestory/ShapesExt.kt
@@ -3,4 +3,9 @@ package ru.acediat.simplestory
 import android.graphics.Rect
 import ru.acediat.engine.shapes.AcRect
 
-fun fromAcRect(rect: AcRect): Rect = Rect(rect.left, rect.top, rect.right, rect.bottom)
+fun fromAcRect(rect: AcRect): Rect = Rect(
+    rect.left.toInt(),
+    rect.top.toInt(),
+    rect.right.toInt(),
+    rect.bottom.toInt()
+)
